requirements.txt
txt 
django
python-dotenv
gunicorn
psycopg2
djangorestframework
djangorestframework-simplejwt
djangorestframework django-filter